+++ app/state/middlewares/logger/index.js
@@ -1,0 +1,7 @@
+export default store => next => action => {
+  console.log('prevState', store.getState())
+  console.log('action', action)
+  const nextAction = next(action)
+  console.log('nextState', store.getState())
+  return nextAction
+}
